create project folder structure
