+++ lib/game/chess_piece.dart
@@ -1,0 +1,20 @@
+class ChessPiece {
+  final PieceType type;
+  final PieceColour colour;
+
+  const ChessPiece(this.type, this.colour);
+}
+
+enum PieceType {
+  king,
+  queen,
+  bishop,
+  knight,
+  rook,
+  pawn,
+}
+
+enum PieceColour {
+  black,
+  white,
+}
